Extract validator mapping helpers
